add bare CompletOrder (#4)
* add bare CompleteOrder test

* add gqlgenc.yaml

+++ internal/handler/http/internal/customer/v1/graphql/request/order.go
@@ -9,7 +9,7 @@ import (
 func NewCreateOrderInput(i graphql.CreateOrderInput) *usecase.CreateOrderInput {
 	// TODO: validate uint or define uint as scaler
 	return &usecase.CreateOrderInput{
-		ProductVariantID: i.ProductVariantID,
+		ProductVariantID: uint(i.ProductVariantID),
 		UnitPrice:        uint(i.UnitPrice),
 		Quantity:         uint(i.Quantity),
 		PaymentMethod:    getPaymentMethod(i.PaymentMethod),
